No need to promote scalars to 64-bit
Now that we're correctly aligning perf-event arguments[1], smaller
scalar values work properly.

[1] Since commit 3630487f35171d4c5e372c4e633fde643f9b45da

+++ tests/codegen/call_printf.cpp
@@ -9,7 +9,7 @@ TEST(codegen, call_printf)
   test("struct Foo { char c; long l; } kprobe:f { $foo = (struct Foo*)0; printf(\"%c %lu\\n\", $foo->c, $foo->l) }",
 
 #if LLVM_VERSION_MAJOR < 7
-R"EXPECTED(%printf_t = type { i64, i64, i64 }
+R"EXPECTED(%printf_t = type { i64, i8, i64 }
 
 ; Function Attrs: nounwind
 declare i64 @llvm.bpf.pseudo(i64, i64) #0
@@ -31,7 +31,7 @@ entry:
   %3 = load i8, i8* %"struct Foo.c", align 1
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %"struct Foo.c")
   %4 = getelementptr inbounds %printf_t, %printf_t* %printf_args, i64 0, i32 1
-  store i8 %3, i64* %4, align 8
+  store i8 %3, i8* %4, align 8
   %5 = bitcast i64* %"struct Foo.l" to i8*
   call void @llvm.lifetime.start.p0i8(i64 -1, i8* nonnull %5)
   %probe_read1 = call i64 inttoptr (i64 4 to i64 (i8*, i64, i8*)*)(i64* nonnull %"struct Foo.l", i64 8, i64 8)
